[feat] : 템플릿 클래스 기본 세팅
