+++ Projects/ABKit/Sources/Template/BaseView.swift
@@ -10,16 +10,24 @@ import UIKit
 
 open class BaseView: UIView {
     
-    open func style() {
-        
+    public init(){
+        super.init(frame: .zero)
+        style()
+        hierarchy()
+        layout()
+        initialize()
     }
     
-    open func hierarchy() {
-        
+    public required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
     }
     
-    open func layout() {
-        
-    }
+    open func style() { }
+    
+    open func hierarchy() { }
+    
+    open func layout() { }
+    
+    open func initialize() { }
     
 }
